requirements.txt
streamlit
google-generative-ai
python-docx
python-dotenv